finished follow feature 	modified:   calendar/app/controllers/Events.java 	modified:   calendar/app/models/Calendar.java 	modified:   calendar/app/models/Event.java 	modified:   calendar/app/views/Calendars/showCalendars.html 	modified:   calendar/app/views/Events/addEvent.html 	modified:   calendar/app/views/Events/editEvent.html 	modified:   calendar/app/views/Events/selectCalendarToFollowEvent.html 	new file:   calendar/app/views/Events/showFollowingEvents.html 	modified:   calendar/app/views/tags/eventsTag.html 	new file:   calendar/app/views/tags/followEventsTag.html 	modified:   calendar/conf/bootstrap-data.yml 	modified:   calendar/conf/routes 	modified:   calendar/test/modelTest/CalendarTest.java 	modified:   calendar/test/modelTest/EventTest.java

+++ calendar/app/controllers/Events.java
@@ -1,5 +1,6 @@
 package controllers;
 
+import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
@@ -23,12 +24,14 @@ public class Events extends Controller {
     public static void createEvent(Long calendarId, @Required String name,
 	    @Required Date startDate, @Required Date endDate,
 	    @Required String startTime, @Required String endTime,
-	    @Required boolean isPublic, String note) {
+	    @Required boolean isPublic, @Required boolean isFollowable,
+	    String note) {
 
 	validation.required(name);
 	validation.required(startDate);
 	validation.required(endDate);
 	validation.required(isPublic);
+	validation.required(isFollowable);
 	validation.match(startTime, regexTime).message("Invalid!");
 	validation.match(endTime, regexTime).message("Invalid!");
 
@@ -44,10 +47,10 @@ public class Events extends Controller {
 	endDate = helperCreateDate(endDate, endTime, "HH:mm");
 
 	User owner = calendar.owner;
-	Event event = new Event(name, note, startDate, endDate, owner,
-		calendar, isPublic, false);
+	new Event(name, note, startDate, endDate, owner, calendar, isPublic,
+		isFollowable).save();
 
-	event.save();
+	calendar.save();
 
 	String nickname = calendar.owner.nickname;
 	Calendars.showCalendar(nickname, calendarId);
@@ -63,10 +66,11 @@ public class Events extends Controller {
 	return aDate.toDate();
     }
 
-    public static void updateEvent(Long eventId, @Required String name,
-	    @Required Date startDate, @Required Date endDate,
-	    @Required String startTime, @Required String endTime,
-	    @Required boolean isPublic, String note) {
+    public static void updateEvent(Long eventId, Long calendarId,
+	    @Required String name, @Required Date startDate,
+	    @Required Date endDate, @Required String startTime,
+	    @Required String endTime, @Required boolean isPublic,
+	    @Required boolean isFollowable, String note) {
 
 	validation.required(name);
 	validation.required(startDate);
@@ -78,7 +82,7 @@ public class Events extends Controller {
 	if (validation.hasErrors()) {
 	    params.flash();
 	    validation.keep();
-	    editEvent(eventId);
+	    editEvent(eventId, calendarId);
 	}
 
 	Event event = Event.findById(eventId);
@@ -89,19 +93,37 @@ public class Events extends Controller {
 	event.upperBound = Event.makeUpperBound(endDate);
 	event.note = note;
 	event.isPublic = isPublic;
+	event.isFollowable = isFollowable;
 
+	/*
+	 * if an event is set to private or not followable, then all user which
+	 * follow this event are set to non followers.
+	 */
+	if (isFollowable == false || isPublic == false) {
+	    List<Calendar> calendars = new ArrayList<Calendar>(event.calendars);
+	    for (Calendar calendar : calendars) {
+		/*
+		 * if the calendar owner is not equals to the event owner, then
+		 * the calendar owner is a follower of this event.
+		 */
+		if (calendar.owner != event.owner) {
+		    event.unfollow(calendar);
+		    event.save();
+		    calendar.save();
+		}
+	    }
+	}
 	event.save();
-
 	String nickname = event.owner.nickname;
 	// TODO:
-	Long calendarId = event.calendars.get(0).id;
-	Calendars.showCalendar(nickname, calendarId);
+	Date aDate = event.start;
+	Calendars.showDate(nickname, calendarId, aDate);
     }
 
-    public static void editEvent(Long eventId) {
+    public static void editEvent(Long eventId, Long calendarId) {
 	Event event = Event.findById(eventId);
 	// TODO:
-	Calendar calendar = event.calendars.get(0);
+	Calendar calendar = Calendar.find("byId", calendarId).first();
 	User user = event.owner;
 	render(calendar, event, user);
     }
@@ -112,28 +134,90 @@ public class Events extends Controller {
 	render(calendar, user);
     }
 
-    public static void removeEvent(Long eventId) {
+    public static void removeEvent(Long eventId, Long calendarId) {
+	Calendar calendar = Calendar.find("byId", calendarId).first();
 	Event event = Event.findById(eventId);
-	event.delete();
+	calendar.events.remove(event);
+	event.calendars.remove(calendar);
+
+	event.save();
+	if (event.calendars.size() == 0) {
+	    event.delete();
+	}
+	calendar.save();
+
 	String nickname = event.owner.nickname;
-	Long calendarId = event.calendars.get(0).id;
 	Calendars.showCalendar(nickname, calendarId);
     }
 
-    // TODO: implement this controller
-    public static void selectCalendarToFollowEvent(Long eventId, Long originalCalendarId) {
-    	User connectedUser = User.find("byNickname", Security.connected()).first();
-    	List<Calendar> calendars =connectedUser.calendars;
-    	Calendar originalCalendar = Calendar.find("byId", originalCalendarId).first();
-    	Event event = Event.find("byID", eventId).first();
-    	render(calendars, originalCalendar, event);
+    public static void showFollowingEvents() {
+	User user = User.find("byNickname", Security.connected()).first();
+
+	List<Calendar> calendars = user.calendars;
+	List<Event> followEvents = new ArrayList<Event>();
+
+	for (Calendar calendar : calendars) {
+	    followEvents.addAll(calendar.getFollowingEvents());
+	}
+
+	render(followEvents);
     }
-    public static void followEvent(Long ownCalendarId, Long originalCalendarId, Long eventId){
-    	String nickname = Security.connected();
-    	Event event = Event.find("byId", eventId).first();
-    	Calendar calendar = Calendar.find("byId", ownCalendarId).first();
-    	event.follow(calendar);
-    	event.save();
-    	Calendars.showCalendar(nickname, originalCalendarId);
+
+    public static void showFollowedEvents() {
+	User user = User.find("byNickname", Security.connected()).first();
+
+	List<Calendar> calendars = user.calendars;
+	List<Event> events = new ArrayList<Event>();
+
+	for (Calendar calendar : calendars) {
+	    events.addAll(calendar.events);
+	}
+	ArrayList<Event> followEvents = new ArrayList<Event>();
+	for (Event ev : events) {
+	    if (ev.isFollowed()) {
+		followEvents.add(ev);
+	    }
+	}
+	render(followEvents);
+    }
+
+    public static void selectCalendarToFollowEvent(Long eventId,
+	    Long originalCalendarId) {
+	User connectedUser = User.find("byNickname", Security.connected())
+		.first();
+	List<Calendar> calendars = connectedUser.calendars;
+	Calendar originalCalendar = Calendar.find("byId", originalCalendarId)
+		.first();
+	Event event = Event.find("byId", eventId).first();
+	render(calendars, originalCalendar, event);
+    }
+
+    public static void followEvent(Long ownCalendarId, Long originalCalendarId,
+	    Long eventId) {
+
+	Event event = Event.find("byId", eventId).first();
+	Calendar calendar = Calendar.find("byId", ownCalendarId).first();
+
+	event.follow(calendar);
+	event.save();
+	calendar.save();
+
+	String nickname = Security.connected();
+	Date aDate = event.start;
+
+	Calendars.showDate(nickname, originalCalendarId, aDate);
+    }
+
+    public static void unfollowEvent(Long calendarId, Long eventId) {
+	Calendar calendar = Calendar.find("byId", calendarId).first();
+	Event event = Event.find("byId", eventId).first();
+
+	event.unfollow(calendar);
+	event.save();
+	calendar.save();
+
+	Date aDate = event.start;
+	String nickname = Security.connected();
+	Calendars.showDate(nickname, calendarId, aDate);
     }
 }
